+++ src/client/bin/main.rs
@@ -4,7 +4,7 @@ use cartel::client::cli::cli_app;
 use cartel::{teprint, texit};
 
 fn main() {
-  if let Err(e) = cli_app() {
-    texit!(e);
-  }
+    if let Err(e) = cli_app() {
+        texit!(e);
+    }
 }