+++ src/client/commands/restart.rs
@@ -1,13 +1,13 @@
-use crate::client::cli::CliOptions;
+use crate::client::cli::ClientConfig;
 use crate::client::emoji::{HOUR_GLASS, SUCCESS};
 use crate::client::request;
 use anyhow::Result;
 use console::style;
 
-pub fn restart_module_cmd(module: &str, cli_config: &CliOptions) -> Result<()> {
+pub fn restart_module_cmd(module: &str, cfg: &ClientConfig) -> Result<()> {
     #[rustfmt::skip]
     tprintstep!(format!("Restarting service '{}'...", module), 1, 2, HOUR_GLASS);
-    request::restart_module(module, &cli_config.daemon_url)?;
+    request::restart_module(module, &cfg.daemon_url)?;
     tprintstep!(style("Service restarted").bold().green(), 2, 2, SUCCESS);
     Ok(())
 }